Restructuring backend code for separation of concerns
