+++ backend/app/models/route.py
@@ -1,0 +1,6 @@
+charging_stations = [
+    {'location': [34.05, -118.25], 'type': 'Fast Charger', 'charging_speed': 120},
+    {'location': [36.16, -115.15], 'type': 'Regular Charger', 'charging_speed': 50},
+    {'location': [37.77, -122.41], 'type': 'Fast Charger', 'charging_speed': 120},
+    {'location': [35.68, -105.94], 'type': 'Regular Charger', 'charging_speed': 50},
+]
